+++ src/Exceptions/PaperException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace JacobJoergensen\LaravelPaper\Exceptions;
+
+use Throwable;
+
+interface PaperException extends Throwable {}